Implement control flow: if and for
This uses a "meta-circular" approach where control-flow structures in
OCaml (the host language) are used to implement control-flow in rhotic
(the interpreted language). I.e. we use branching and iteration, and
then recursively call the interpreter.

(Not yet implemented: functions. The approach should be similar: create
a new environment, binding the arguments to the parameters, and then
recursively calling eval.)

This approach is slightly easier to implement, but the (interpreted) program
is implicit. The calls are also not tail recursive.

With a little work, recursion could be removed by maintaining a program
state with a list of statements to execute; branching or looping is
handled by prepending statements to the list.

Yet another approach is to make more of the program state, including the
program counter, explicit. This approach requires a "lowering" step, to
flatten all nested statements into a single array of statements (where
the program counter is an index into the array), and control flow is
compiled into labels and jumps.

For now, we use the simpler approach, but if needed, we can switch to
the explicit program state approach.