Add an example for a more complicated monitor
We want the monitor to track and propagate constraints, e.g. the types
of variables, and also whether NAs are allowed.

In this example, `cond` is used by the `if` statement; therefore `cond`
has to be boolean-like and cannot be an NA. If we propagate these
constraints backwards, we find that `mod` cannot be NA, and `i` cannot
be NA, therefore `x` cannot contain NAs. We also observe that `cond` has
to be a boolean (not just boolean-like) because it is the result of a
comparison.