Better handling of top-level "inputs."
Before, we had a hack to handle top-level assignments properly. This was
so that constant assignments at the top level could be treated as
"inputs," otherwise there would be no dependencies as all variables are
constant.

Now, we can "tag" assignments, e.g. "x <- .input(v)" assigns v to x, but
marks x as an input. I.e. we make x a self-dependency, so that other
variables can depend on x.